fix bug model food

+++ app/Models/Food.php
@@ -27,7 +27,7 @@ class Food extends Model {
      * @var array
      */
     protected $casts = [
-        'rate' => 'integer',
+        'rate' => 'double',
         'price' => 'integer',
         'quantity' => 'integer',
         'discount' => 'integer',
